Use cron resource to trigger link generation
This commit updates the resource which triggers the link generation to use a `cron-resource` rather than `time`. The purpose of this is to ensure that the pipeline isn't triggered at random times when the underlying Concourse instance is restarted or upgraded, as the `time` resource loses its state when this happens and triggers the pipeline when it next checks if it should be run. The frequency is also changed from every three weeks to every two weeks.