Fix destroy instance steps
This commit fixes the destroy instance steps to reference the correct steps per the environment. Previously, we had created a YAML anchor for `destroy-ingestion-instance` (defined for integration) and referenced this from the staging and production steps, however this meant we were referencing the incorrect Concourse role ARN for those steps (although practically it is the same at the moment).