Add missing resource type for cron-resource
Recently the `cron-resource` was added to the pipeline to control when the generation of related links is kicked off (replacing the `time` resource). As the `cron-resource` is not bundled with Concourse by default, we need to add a custom `resource-type` for it to work, which is what this commit retrospectively does.