Allow unset variables in generation provisioning script
This commit updates the `provision_generation_machine` script to allow unset variables, without treating them as an error and exiting immediately as is otherwise set by `set -u`.

The reason this change is necessary is because the generation pipeline has recently started to fail (without any changes to the pipeline itself). Investigating this further, the `run_link_generation` script isn't being run as the provisioning script never completes; it appears that there is an unset variable when running `Processing triggers for libc-bin (2.27-3ubuntu1)` (part of `Setting up build-essential (12.4ubuntu1)`). This is likely to be as a recent change of this or a dependent package.

Allowing unset variables fixes this problem.