Delegate Kernel.sleep to instance method to allow testing
This commit adds a new method `wait` to `RelatedLinksUpdater` which simple calls `Kernel.sleep`. By removing the call to `Kernel.sleep` during testing, we do not waste time waiting for this call to evaluate (as it does real sleeping). This also removes complications stubbing `Kernel`, due to the fact that the `sleep` method is inherited from `Kernel` but isn't actually called from the `Kernel` class object.