Remove pointless sleep on last ingestion batch
This commit removes a pointless sleep on the last ingestion batch in the case where the number of updates exactly equals the batch size. For example, if we previously had a batch size of 10 updates per batch, and we had a single batch of 10 updates, we would wait for `between_batch_wait_time_seconds` before recommencing processing. This logic was used to determine when we should sleep, however it had this edge case which added one additional sleep, prolonging the time taken to complete ingestion.

It has been fixed in this commit by comparing the index of the current batch with the index of the last batch, to ensure we sleep between every batch and not after the last batch.