Change functional network window
This commit changes the functional network window (i.e. the period for which we obtain user movement data) so that we use data up to two days ago (changed from the previous day). The reason for this change is that Google moves data from a Google Analytics view into a specific BigQuery table on a daily basis, and this happens in the early hours of the day pacific standard time. As a result, we have no guarantee that the data we need in BigQuery is present when running the related links pipeline at 8am.

The change in this commit addresses this problem by generating data for two days ago from the current date, which is guaranteed to be in BigQuery.