Freeze all dependencies
A bug appeared when pip upgraded protobuf (who was set by google-cloud-bigquery-1.17.0
to be >=3.6.0) to 4.21.0, which requires python >= 3.7, while the version here is 3.6.
So we freeze the version back to 3.19.4, as it was before, in order to resolve the
dependencies again.
And we then just freeze all packages in requirements.txt, which should have been the case
already. The previous requirements file is now called requirements-original.txt, in case
the frozen requirements should need to be regenerated