Round off confidence scores to ignore floating point errors
This is to address tests failing randomly.

Given the random process used to compute confidence scores, they can slightly
vary with no consequence on the final output. The related links tests aren't meant
to check that, so we round off the confidence scores, so that it's the actual related
links result that is compared.